feat(gallery): move JourneyItem to Gallery.kt and support image
- Move JourneyItem from Overview.kt to Gallery.kt.
- Implement image support using rememberAsyncImagePainter for journey.imageUrl.
